Move utility code for sort filter into its own file.
